+++ test/object-store/sync/flx_sync.cpp
@@ -3797,7 +3797,6 @@ TEST_CASE("flx: open realm + register subscription callack while bootstrapping",
     }
 
     SECTION("Sync Open + Async Open") {
-
         {
             subscription_invoked = false;
             config.sync_config->subscription_initializer = init_subscription_callback;
@@ -3813,7 +3812,7 @@ TEST_CASE("flx: open realm + register subscription callack while bootstrapping",
         {
             auto subscription_pf_async = util::make_promise_future<bool>();
             auto init_subscription_asyc_callback =
-                [&, promise_holder_async = util::CopyablePromiseHolder(std::move(subscription_pf_async.promise))](
+                [promise_holder_async = util::CopyablePromiseHolder(std::move(subscription_pf_async.promise))](
                     std::shared_ptr<Realm> realm) mutable {
                     REQUIRE(realm);
                     auto table = realm->read_group().get_table("class_TopLevel");
@@ -3823,7 +3822,6 @@ TEST_CASE("flx: open realm + register subscription callack while bootstrapping",
                     mutable_subscription.insert_or_assign(query);
                     auto promise = promise_holder_async.get_promise();
                     mutable_subscription.commit();
-                    subscription_invoked = true;
                     promise.emplace_value(true);
                 };
             auto open_realm_pf = util::make_promise_future<bool>();
@@ -3843,6 +3841,10 @@ TEST_CASE("flx: open realm + register subscription callack while bootstrapping",
             async_open->start(open_realm_completed_callback);
             REQUIRE(open_realm_pf.future.get());
             REQUIRE(subscription_pf_async.future.get());
+            config.sync_config->rerun_init_subscription_on_open = false;
+            auto realm = Realm::get_shared_realm(config);
+            REQUIRE(realm->get_latest_subscription_set().version() == 2);
+            REQUIRE(realm->get_active_subscription_set().version() == 2);
         }
     }
 
@@ -3902,66 +3904,88 @@ TEST_CASE("flx: open realm + register subscription callack while bootstrapping",
         }
 
         SECTION("rerun on open set for multiple async open tasks (subscription runs only once)") {
-            auto subscription_pf = util::make_promise_future<bool>();
-            std::atomic<int> cnt = 0;
-            auto init_subscription_callback =
-                [&, promise_holder = util::CopyablePromiseHolder(std::move(subscription_pf.promise))](
-                    std::shared_ptr<Realm> realm) mutable {
-                    cnt += 1;
-                    REQUIRE(cnt == 1);
-                    REQUIRE(realm);
-                    auto table = realm->read_group().get_table("class_TopLevel");
-                    Query query(table);
-                    auto subscription = realm->get_latest_subscription_set();
-                    auto mutable_subscription = subscription.make_mutable_copy();
-                    mutable_subscription.insert_or_assign(query);
-                    mutable_subscription.commit();
-                    promise_holder.get_promise().emplace_value(true);
+
+            std::atomic<int> init_sub_cnt = 0;
+            auto subscription_pf_task1 = util::make_promise_future<bool>();
+            auto subscription_pf_task2 = util::make_promise_future<bool>();
+
+            auto init_subscription = [&](auto&& subscription_pf, std::shared_ptr<Realm> realm) {
+                realm->refresh();
+                auto promise_holder = util::CopyablePromiseHolder(std::move(subscription_pf.promise));
+                init_sub_cnt += 1;
+                REQUIRE(realm);
+                auto table = realm->read_group().get_table("class_TopLevel");
+                Query query(table);
+                auto subscription = realm->get_latest_subscription_set();
+                auto mutable_subscription = subscription.make_mutable_copy();
+                mutable_subscription.insert_or_assign(query);
+                mutable_subscription.commit();
+                promise_holder.get_promise().emplace_value(true);
+            };
+
+            auto async_open_realm = [&init_subscription, &config](auto&& subscription_pf) {
+                // commit something in the database
+
+                config.sync_config->subscription_initializer = [&](std::shared_ptr<Realm> realm) {
+                    init_subscription(subscription_pf, realm);
                 };
-            config.sync_config->subscription_initializer = init_subscription_callback;
-            config.sync_config->rerun_init_subscription_on_open = true;
+                config.sync_config->rerun_init_subscription_on_open = true;
+                auto async_open_task = Realm::get_synchronized_realm(config);
 
-            auto async_open_task1 = Realm::get_synchronized_realm(config);
+                auto open_task_pf = util::make_promise_future<SharedRealm>();
+                auto open_callback = [promise_holder = util::CopyablePromiseHolder(std::move(open_task_pf.promise))](
+                                         ThreadSafeReference ref, std::exception_ptr err) mutable {
+                    REQUIRE_FALSE(err);
+                    auto realm = Realm::get_shared_realm(std::move(ref));
+                    REQUIRE(realm);
+                    promise_holder.get_promise().emplace_value(realm);
+                };
 
-            // I need to investigate this further. Since there is a chance that:
-            // task 1 runs after task2 and it is constructed with m_first_db_open == true
-            // task 2 is constructed with m_first_db_open == false but the transaction version is 0
-            // both tasks now will run the subscription callback, which does not seem correct.
-
-            config.sync_config->rerun_init_subscription_on_open = false;
-            auto async_open_task2 = Realm::get_synchronized_realm(config);
-
-            auto open_t1_pf = util::make_promise_future<SharedRealm>();
-            auto open_t2_pf = util::make_promise_future<SharedRealm>();
-
-            auto open_callback_task_1 = [&,
-                                         promise_holder = util::CopyablePromiseHolder(std::move(open_t1_pf.promise))](
-                                            ThreadSafeReference ref, std::exception_ptr err) mutable {
-                REQUIRE_FALSE(err);
-                auto realm = Realm::get_shared_realm(std::move(ref));
-                REQUIRE(realm);
-                promise_holder.get_promise().emplace_value(realm);
-            };
-            auto open_callback_task_2 = [&,
-                                         promise_holder = util::CopyablePromiseHolder(std::move(open_t2_pf.promise))](
-                                            ThreadSafeReference ref, std::exception_ptr err) mutable {
-                REQUIRE_FALSE(err);
-                auto realm = Realm::get_shared_realm(std::move(ref));
-                REQUIRE(realm);
-                promise_holder.get_promise().emplace_value(realm);
+                async_open_task->start(open_callback);
+                return open_task_pf;
             };
 
-            async_open_task1->start(open_callback_task_1);
-            async_open_task2->start(open_callback_task_2);
+            SECTION("Realm was already created") {
+                subscription_invoked = false;
+                config.sync_config->subscription_initializer = init_subscription_callback;
+                auto realm = Realm::get_shared_realm(config);
+                REQUIRE(subscription_pf.future.get());
+                auto sb = realm->get_latest_subscription_set();
+                auto future = sb.get_state_change_notification(realm::sync::SubscriptionSet::State::Complete);
+                auto state = future.get();
+                REQUIRE(state == realm::sync::SubscriptionSet::State::Complete);
+                realm->refresh(); // refresh is needed otherwise table_ref->size() would be 0
+                REQUIRE(verify_subscription(realm));
+                REQUIRE(realm->get_latest_subscription_set().version() == 1);
+                REQUIRE(realm->get_active_subscription_set().version() == 1);
 
-            // subscription init called only once but realm opened twice
-            auto realm1 = open_t1_pf.future.get();
-            auto realm2 = open_t2_pf.future.get();
-            REQUIRE(realm1->get_latest_subscription_set().version() == 1);
-            REQUIRE(realm1->get_active_subscription_set().version() == 1);
-            REQUIRE(realm2->get_latest_subscription_set().version() == 1);
-            REQUIRE(realm2->get_active_subscription_set().version() == 1);
-            REQUIRE(subscription_pf.future.get());
+                auto open_callback_task1 = async_open_realm(subscription_pf_task1);
+                auto open_callback_task2 = async_open_realm(subscription_pf_task2);
+                auto realm1 = open_callback_task1.future.get();
+                auto realm2 = open_callback_task2.future.get();
+                REQUIRE(init_sub_cnt == 1); // subscription callback invoked only once
+                REQUIRE(realm1->get_latest_subscription_set().version() == 2);
+                REQUIRE(realm1->get_active_subscription_set().version() == 2);
+                REQUIRE(realm2->get_latest_subscription_set().version() == 2);
+                REQUIRE(realm2->get_active_subscription_set().version() == 2);
+            }
+            SECTION("First time realm is created but opened via open async") {
+
+                auto open_callback_task1 = async_open_realm(subscription_pf_task1);
+                auto open_callback_task2 = async_open_realm(subscription_pf_task2);
+                // open
+                auto realm1 = open_callback_task1.future.get();
+                auto realm2 = open_callback_task2.future.get();
+                REQUIRE(subscription_pf.future.get());
+                REQUIRE(init_sub_cnt.load() == 2);
+                // we may have run the subscription init callback 2 times, this can only happen
+                // if this is the first time we have created a realm and we have launched 2 async open
+                // tasks in parallel.
+                REQUIRE(realm1->get_latest_subscription_set().version() >= 2);
+                REQUIRE(realm1->get_active_subscription_set().version() >= 2);
+                REQUIRE(realm2->get_latest_subscription_set().version() >= 2);
+                REQUIRE(realm2->get_active_subscription_set().version() >= 2);
+            }
         }
     }
 }